Split out sub-command options definition
Sub-command options are now defined in dedicated files:
  commands/aws.py -> AWS
  commands/rds.py -> AWS RDS
  commands/aurora.py -> AWS RDS Aurora
  commands/gcloud.py -> GCloud
  commands/azure.py -> Azure
