+++ edbdeploy/cli.py
@@ -1,195 +1,15 @@
 import argcomplete
 import argparse
-import os
-import re
 import sys
-import textwrap
 
 from . import __version__
-from .project import Project
-
-class ReferenceArchitectureOption:
-    choices = ['EDB-RA-1', 'EDB-RA-2', 'EDB-RA-3', 'HammerDB-DBaaS',
-               'HammerDB-TPROC-C']
-
-    default = 'EDB-RA-1'
-    help = textwrap.dedent("""
-        Reference architecture code name. Allowed values are: EDB-RA-1 for a
-        single Postgres node deployment with one backup server and one PEM
-        monitoring server, EDB-RA-2 for a 3 Postgres nodes deployment with
-        quorum base synchronous replication and automatic failover, one backup
-        server and one PEM monitoring server, EDB-RA-3 for extending EDB-RA-2
-        with 3 PgPoolII nodes, and HammerDB-TPROC-C for benchmarking 2-tier
-        client-server architectures with and OLTP workload. Default:
-        %(default)s
-    """)
-
-
-class ReferenceArchitectureOptionRDS:
-    choices = ['HammerDB-DBaaS']
-    default = 'HammerDB-DBaaS'
-    help = textwrap.dedent("""
-        Reference architecture code name. Allowed values are: HammerDB-DBaaS
-        for benchmarking any DBaaS offering for benchmarking 2-tier
-        client-server architectures with and OLTP workload. Default:
-        %(default)s
-    """)
-
-
-class OSOption:
-    choices = ['CentOS7', 'CentOS8', 'RedHat7', 'RedHat8']
-    default = 'CentOS8'
-    help = textwrap.dedent("""
-        Operating system. Allowed values are: CentOS7, CentOS8, RedHat7 and
-        RedHat8. Default: %(default)s
-    """)
-
-
-class PgVersionOption:
-    choices = ['11', '12', '13']
-    default = '13'
-    help = textwrap.dedent("""
-        PostgreSQL or EPAS version. Allowed values are: 11, 12 and 13.
-        Default: %(default)s
-    """)
-
-
-class EFMVersionOption:
-    choices = ['3.10', '4.0', '4.1']
-    default = '4.1'
-    help = textwrap.dedent("""
-        EDB Failover Manager version. Allowed values are: 3.10, 4.0 and 4.1.
-        Default: %(default)s
-    """)
-
-
-class PgTypeOption:
-    choices = ['PG', 'EPAS']
-    default = 'PG'
-    help = textwrap.dedent("""
-        Postgres engine type. Allowed values are: PG for PostgreSQL, EPAS for
-        EDB Postgres Advanced Server. Default:
-        %(default)s
-    """)
-
-
-class PgTypeOptionRDS:
-    choices = ['RDS']
-    default = 'RDS'
-    help = textwrap.dedent("""
-        Postgres engine type. Allowed values are: RDS for AWS RDS.  Default:
-        %(default)s
-    """)
-
-
-class SSHPubKeyOption:
-
-    help = textwrap.dedent("""
-        SSH public key path to use. Default: %(default)s
-    """)
-
-    @staticmethod
-    def default():
-        home = os.path.expanduser("~")
-        pub_key_path = os.path.join(home, '.ssh', 'id_rsa.pub')
-        if os.path.exists(pub_key_path):
-            return pub_key_path
-
-
-class SSHPrivKeyOption:
-
-    help = textwrap.dedent("""
-        SSH private key path to use. Default: %(default)s
-    """)
-
-    @staticmethod
-    def default():
-        home = os.path.expanduser("~")
-        priv_key_path = os.path.join(home, '.ssh', 'id_rsa')
-        if os.path.exists(priv_key_path):
-            return priv_key_path
-
-
-# HammerDB specific options
-
-class ShirtSizeOption:
-    choices = ['small', 'medium', 'large', 'xl']
-    default = 'small'
-    help = textwrap.dedent("""
-        T-shirt sized system to provision.  Allowed values are small, medium,
-        large, and xl.  Default : %(default)s
-    """)
-
-
-# Cloud specific options
-class AWSRegionOption:
-    choices = ['us-east-1', 'us-east-2', 'us-west-1', 'us-west-2']
-    default = 'us-east-1'
-    help = textwrap.dedent("""
-        AWS region. Allowed values are us-east-1, us-east-2, us-west-1 and
-        us-west-2. Default: %(default)s
-    """)
-
-
-class AWSIAMIDOption:
-    default = ''
-    help = textwrap.dedent("""
-        AWS Image ID. Default: %(default)s
-    """)
-
-
-class AzureRegionOption:
-    choices = ['centralus', 'eastus', 'eastus2', 'westus', 'westcentralus',
-               'westus2', 'northcentralus', 'southcentralus']
-    default = 'eastus'
-    help = textwrap.dedent("""
-        Azure region. Allowed values are centralus, eastus, eastus2, westus,
-        westcentralus, westus2, northcentralus and southcentralus.
-        Default: %(default)s
-    """)
-
-
-class GCloudRegionOption:
-    choices = ['us-central1', 'us-east1', 'us-east4', 'us-west1', 'us-west2',
-               'us-west3', 'us-west4']
-    default = 'us-east1'
-    help = textwrap.dedent("""
-        GCloud region. Allowed values are us-central1, us-east1, us-east4,
-        us-west1, us-west2, us-west3 and us-west4. Default: %(default)s
-    """)
-
-
-class GCloudCredentialsOption:
-
-    help = textwrap.dedent("""
-        GCloud credentials file (JSON) to use. Default: %(default)s
-    """)
-
-    @staticmethod
-    def default():
-        home = os.path.expanduser("~")
-        credential_file = os.path.join(home, 'accounts.json')
-        if os.path.exists(credential_file):
-            return credential_file
-
-
-def EDBCredentialsType(value):
-    p = re.compile(r"^([^:]+):(.+)$")
-    if not p.match(value):
-        raise argparse.ArgumentTypeError(
-            "EDB Credentials does not match \"<username>:<password>\""
-        )
-    return value
-
-
-def ProjectType(value):
-    p = re.compile(r"^[a-z0-9]{3,12}$")
-    if not p.match(value):
-        raise argparse.ArgumentTypeError(
-            "Project name should only contain lower alphanumeric characters, "
-            "length must be between 3 and 12"
-        )
-    return value
+from .commands import (
+    aurora as aws_rds_aurora,
+    aws,
+    azure,
+    rds as aws_rds,
+    gcloud,
+)
 
 
 class CLIParser(argparse.ArgumentParser):
@@ -199,927 +19,6 @@ class CLIParser(argparse.ArgumentParser):
         sys.stderr.write('error: %s\n' % message)
         sys.exit(2)
 
-
-# AWS sub-commands and options
-def aws_subcommands(aws_subparser):
-    aws_configure = aws_subparser.add_parser(
-        'configure', help='Project configuration'
-    )
-    aws_provision = aws_subparser.add_parser(
-        'provision', help='Machines provisioning'
-    )
-    aws_destroy = aws_subparser.add_parser(
-        'destroy', help='Machines destruction'
-    )
-    aws_deploy = aws_subparser.add_parser(
-        'deploy', help='Postgres deployment'
-    )
-    aws_show = aws_subparser.add_parser(
-        'show', help='Show configuration'
-    )
-    aws_display = aws_subparser.add_parser(
-        'display', help='Display project details'
-    )
-    aws_passwords = aws_subparser.add_parser(
-        'passwords', help='Display project passwords'
-    )
-    aws_list = aws_subparser.add_parser(
-        'list', help='List projects'
-    )
-    aws_specs = aws_subparser.add_parser(
-        'specs', help='Show Cloud default specifications'
-    )
-    aws_logs = aws_subparser.add_parser(
-        'logs', help='Show project logs'
-    )
-    aws_remove = aws_subparser.add_parser(
-        'remove', help='Remove project'
-    )
-    # aws configure sub-command options
-    aws_configure.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    aws_configure.add_argument(
-        '-a', '--reference-architecture',
-        dest='reference_architecture',
-        choices=ReferenceArchitectureOption.choices,
-        default=ReferenceArchitectureOption.default,
-        metavar='<ref-arch-code>',
-        help=ReferenceArchitectureOption.help
-    )
-    aws_configure.add_argument(
-        '-u', '--edb-credentials',
-        dest='edb_credentials',
-        required=True,
-        type=EDBCredentialsType,
-        metavar='"<username>:<password>"',
-        help="EDB Packages repository credentials."
-    ).completer = edb_credentials_completer
-    aws_configure.add_argument(
-        '-o', '--os',
-        dest='operating_system',
-        choices=OSOption.choices,
-        default=OSOption.default,
-        metavar='<operating-system>',
-        help=OSOption.help
-    )
-    aws_configure.add_argument(
-        '-t', '--pg-type',
-        dest='postgres_type',
-        choices=PgTypeOption.choices,
-        default=PgTypeOption.default,
-        metavar='<postgres-engine-type>',
-        help=PgTypeOption.help
-    )
-    aws_configure.add_argument(
-        '-v', '--pg-version',
-        dest='postgres_version',
-        choices=PgVersionOption.choices,
-        default=PgVersionOption.default,
-        metavar='<postgres-version>',
-        help=PgVersionOption.help
-    )
-    aws_configure.add_argument(
-        '-e', '--efm-version',
-        dest='efm_version',
-        choices=EFMVersionOption.choices,
-        default=EFMVersionOption.default,
-        metavar='<efm-version>',
-        help=EFMVersionOption.help
-    )
-    aws_configure.add_argument(
-        '-k', '--ssh-pub-key',
-        dest='ssh_pub_key',
-        type=argparse.FileType('r'),
-        default=SSHPubKeyOption.default(),
-        metavar='<ssh-public-key-file>',
-        help=SSHPubKeyOption.help
-    )
-    aws_configure.add_argument(
-        '-K', '--ssh-private-key',
-        dest='ssh_priv_key',
-        type=argparse.FileType('r'),
-        default=SSHPrivKeyOption.default(),
-        metavar='<ssh-private-key-file>',
-        help=SSHPrivKeyOption.help
-    )
-    aws_configure.add_argument(
-        '-r', '--aws-region',
-        dest='aws_region',
-        choices=AWSRegionOption.choices,
-        default=AWSRegionOption.default,
-        metavar='<cloud-region>',
-        help=AWSRegionOption.help
-    )
-    aws_configure.add_argument(
-        '-i', '--aws-ami-id',
-        dest='aws_ami_id',
-        type=str,
-        default=AWSIAMIDOption.default,
-        metavar='<aws-ami-id>',
-        help=AWSIAMIDOption.help
-    ).completer = aws_ami_id_completer
-    aws_configure.add_argument(
-        '-s', '--spec',
-        dest='spec_file',
-        type=argparse.FileType('r'),
-        metavar='<aws-spec-file>',
-        help="AWS instances specification file, in JSON."
-    )
-    # aws logs sub-command options
-    aws_logs.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    aws_logs.add_argument(
-        '-t', '--tail',
-        dest='tail',
-        action='store_true',
-        help="Do not stop at the end of file."
-    )
-    # aws remove sub-command options
-    aws_remove.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # aws show sub-command options
-    aws_show.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # aws display sub-command option
-    aws_display.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # aws password sub-command option
-    aws_passwords.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # aws provision sub-command options
-    aws_provision.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # aws destroy sub-command options
-    aws_destroy.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # aws deploy sub-command options
-    aws_deploy.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    aws_deploy.add_argument(
-        '-n', '--no-install-collection',
-        dest='no_install_collection',
-        action='store_true',
-        help="Do not install the Ansible collection."
-    )
-
-# AWS RDS for sub-commands and options
-def aws_rds_subcommands(aws_rds_subparser):
-    aws_rds_configure = aws_rds_subparser.add_parser(
-        'configure', help='Project configuration'
-    )
-    aws_rds_provision = aws_rds_subparser.add_parser(
-        'provision', help='Machines provisioning'
-    )
-    aws_rds_destroy = aws_rds_subparser.add_parser(
-        'destroy', help='Machines destruction'
-    )
-    aws_rds_deploy = aws_rds_subparser.add_parser(
-        'deploy', help='Postgres deployment'
-    )
-    aws_rds_show = aws_rds_subparser.add_parser(
-        'show', help='Show configuration'
-    )
-    aws_rds_display = aws_rds_subparser.add_parser(
-        'display', help='Display project details'
-    )
-    aws_rds_passwords = aws_rds_subparser.add_parser(
-        'passwords', help='Display project password'
-    )
-    aws_rds_list = aws_rds_subparser.add_parser(
-        'list', help='List projects'
-    )
-    aws_rds_specs = aws_rds_subparser.add_parser(
-        'specs', help='Show Cloud default specifications'
-    )
-    aws_rds_logs = aws_rds_subparser.add_parser(
-        'logs', help='Show project logs'
-    )
-    aws_rds_remove = aws_rds_subparser.add_parser(
-        'remove', help='Remove project'
-    )
-    # aws-rds configure sub-command options
-    aws_rds_configure.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    aws_rds_configure.add_argument(
-        '-a', '--reference-architecture',
-        dest='reference_architecture',
-        choices=ReferenceArchitectureOptionRDS.choices,
-        default=ReferenceArchitectureOptionRDS.default,
-        metavar='<ref-arch-code>',
-        help=ReferenceArchitectureOptionRDS.help
-    )
-    aws_rds_configure.add_argument(
-        '-u', '--edb-credentials',
-        dest='edb_credentials',
-        required=True,
-        type=EDBCredentialsType,
-        metavar='"<username>:<password>"',
-        help="EDB Packages repository credentials."
-    ).completer = edb_credentials_completer
-    aws_rds_configure.add_argument(
-        '-o', '--os',
-        dest='operating_system',
-        choices=OSOption.choices,
-        default=OSOption.default,
-        metavar='<operating-system>',
-        help=OSOption.help
-    )
-    aws_rds_configure.add_argument(
-        '-t', '--pg-type',
-        dest='postgres_type',
-        choices=PgTypeOptionRDS.choices,
-        default=PgTypeOptionRDS.default,
-        metavar='<postgres-engine-type>',
-        help=PgTypeOptionRDS.help
-    )
-    aws_rds_configure.add_argument(
-        '-v', '--pg-version',
-        dest='postgres_version',
-        choices=PgVersionOption.choices,
-        default=PgVersionOption.default,
-        metavar='<postgres-version>',
-        help=PgVersionOption.help
-    )
-    aws_rds_configure.add_argument(
-        '-k', '--ssh-pub-key',
-        dest='ssh_pub_key',
-        type=argparse.FileType('r'),
-        default=SSHPubKeyOption.default(),
-        metavar='<ssh-public-key-file>',
-        help=SSHPubKeyOption.help
-    )
-    aws_rds_configure.add_argument(
-        '-K', '--ssh-private-key',
-        dest='ssh_priv_key',
-        type=argparse.FileType('r'),
-        default=SSHPrivKeyOption.default(),
-        metavar='<ssh-private-key-file>',
-        help=SSHPrivKeyOption.help
-    )
-    aws_rds_configure.add_argument(
-        '-r', '--aws-rds-region',
-        dest='aws_region',
-        choices=AWSRegionOption.choices,
-        default=AWSRegionOption.default,
-        metavar='<cloud-region>',
-        help=AWSRegionOption.help
-    )
-    aws_rds_configure.add_argument(
-        '-i', '--aws-rds-ami-id',
-        dest='aws_ami_id',
-        type=str,
-        default=AWSIAMIDOption.default,
-        metavar='<aws-rds-ami-id>',
-        help=AWSIAMIDOption.help
-    ).completer = aws_ami_id_completer
-    aws_rds_configure.add_argument(
-        '-s', '--spec',
-        dest='spec_file',
-        type=argparse.FileType('r'),
-        metavar='<aws-rds-spec-file>',
-        help="AWS instances specification file, in JSON."
-    )
-    aws_rds_configure.add_argument(
-        '-T', '--t-shirt',
-        dest='shirt',
-        choices=ShirtSizeOption.choices,
-        default=ShirtSizeOption.default,
-        metavar='<shirt-size>',
-        help=ShirtSizeOption.help
-    )
-    # aws-rds logs sub-command options
-    aws_rds_logs.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    aws_rds_logs.add_argument(
-        '-t', '--tail',
-        dest='tail',
-        action='store_true',
-        help="Do not stop at the end of file."
-    )
-    # aws-rds remove sub-command options
-    aws_rds_remove.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # aws-rds show sub-command options
-    aws_rds_show.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # aws-rds display sub-command option
-    aws_rds_display.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # aws-rds passwords sub-command option
-    aws_rds_passwords.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # aws-rds provision sub-command options
-    aws_rds_provision.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # aws-rds destroy sub-command options
-    aws_rds_destroy.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # aws-rds deploy sub-command options
-    aws_rds_deploy.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    aws_rds_deploy.add_argument(
-        '-n', '--no-install-collection',
-        dest='no_install_collection',
-        action='store_true',
-        help="Do not install the Ansible collection."
-    )
-
-# AWS RDS Aurora for sub-commands and options
-def aws_rds_aurora_subcommands(aws_rds_aurora_subparser):
-    aws_rds_aurora_configure = aws_rds_aurora_subparser.add_parser(
-        'configure', help='Project configuration'
-    )
-    aws_rds_aurora_provision = aws_rds_aurora_subparser.add_parser(
-        'provision', help='Machines provisioning'
-    )
-    aws_rds_aurora_destroy = aws_rds_aurora_subparser.add_parser(
-        'destroy', help='Machines destruction'
-    )
-    aws_rds_aurora_deploy = aws_rds_aurora_subparser.add_parser(
-        'deploy', help='Postgres deployment'
-    )
-    aws_rds_aurora_show = aws_rds_aurora_subparser.add_parser(
-        'show', help='Show configuration'
-    )
-    aws_rds_aurora_display = aws_rds_aurora_subparser.add_parser(
-        'display', help='Display project details'
-    )
-    aws_rds_aurora_passwords = aws_rds_aurora_subparser.add_parser(
-        'passwords', help='Display project passwords'
-    )
-    aws_rds_aurora_list = aws_rds_aurora_subparser.add_parser(
-        'list', help='List projects'
-    )
-    aws_rds_aurora_specs = aws_rds_aurora_subparser.add_parser(
-        'specs', help='Show Cloud default specifications'
-    )
-    aws_rds_aurora_logs = aws_rds_aurora_subparser.add_parser(
-        'logs', help='Show project logs'
-    )
-    aws_rds_aurora_remove = aws_rds_aurora_subparser.add_parser(
-        'remove', help='Remove project'
-    )
-    # aws-rds-aurora configure sub-command options
-    aws_rds_aurora_configure.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    aws_rds_aurora_configure.add_argument(
-        '-a', '--reference-architecture',
-        dest='reference_architecture',
-        choices=ReferenceArchitectureOptionRDS.choices,
-        default=ReferenceArchitectureOptionRDS.default,
-        metavar='<ref-arch-code>',
-        help=ReferenceArchitectureOptionRDS.help
-    )
-    aws_rds_aurora_configure.add_argument(
-        '-u', '--edb-credentials',
-        dest='edb_credentials',
-        required=True,
-        type=EDBCredentialsType,
-        metavar='"<username>:<password>"',
-        help="EDB Packages repository credentials."
-    ).completer = edb_credentials_completer
-    aws_rds_aurora_configure.add_argument(
-        '-o', '--os',
-        dest='operating_system',
-        choices=OSOption.choices,
-        default=OSOption.default,
-        metavar='<operating-system>',
-        help=OSOption.help
-    )
-    aws_rds_aurora_configure.add_argument(
-        '-t', '--pg-type',
-        dest='postgres_type',
-        choices=PgTypeOptionRDS.choices,
-        default=PgTypeOptionRDS.default,
-        metavar='<postgres-engine-type>',
-        help=PgTypeOptionRDS.help
-    )
-    aws_rds_aurora_configure.add_argument(
-        '-v', '--pg-version',
-        dest='postgres_version',
-        choices=PgVersionOption.choices,
-        default=PgVersionOption.default,
-        metavar='<postgres-version>',
-        help=PgVersionOption.help
-    )
-    aws_rds_aurora_configure.add_argument(
-        '-k', '--ssh-pub-key',
-        dest='ssh_pub_key',
-        type=argparse.FileType('r'),
-        default=SSHPubKeyOption.default(),
-        metavar='<ssh-public-key-file>',
-        help=SSHPubKeyOption.help
-    )
-    aws_rds_aurora_configure.add_argument(
-        '-K', '--ssh-private-key',
-        dest='ssh_priv_key',
-        type=argparse.FileType('r'),
-        default=SSHPrivKeyOption.default(),
-        metavar='<ssh-private-key-file>',
-        help=SSHPrivKeyOption.help
-    )
-    aws_rds_aurora_configure.add_argument(
-        '-r', '--aws-rds-region',
-        dest='aws_region',
-        choices=AWSRegionOption.choices,
-        default=AWSRegionOption.default,
-        metavar='<cloud-region>',
-        help=AWSRegionOption.help
-    )
-    aws_rds_aurora_configure.add_argument(
-        '-i', '--aws-rds-ami-id',
-        dest='aws_ami_id',
-        type=str,
-        default=AWSIAMIDOption.default,
-        metavar='<aws-rds-ami-id>',
-        help=AWSIAMIDOption.help
-    ).completer = aws_ami_id_completer
-    aws_rds_aurora_configure.add_argument(
-        '-s', '--spec',
-        dest='spec_file',
-        type=argparse.FileType('r'),
-        metavar='<aws-rds-spec-file>',
-        help="AWS instances specification file, in JSON."
-    )
-    aws_rds_aurora_configure.add_argument(
-        '-T', '--t-shirt',
-        dest='shirt',
-        choices=ShirtSizeOption.choices,
-        default=ShirtSizeOption.default,
-        metavar='<shirt-size>',
-        help=ShirtSizeOption.help
-    )
-    # aws-rds-aurora logs sub-command options
-    aws_rds_aurora_logs.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    aws_rds_aurora_logs.add_argument(
-        '-t', '--tail',
-        dest='tail',
-        action='store_true',
-        help="Do not stop at the end of file."
-    )
-    # aws-rds remove sub-command options
-    aws_rds_aurora_remove.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # aws-rds-aurora show sub-command options
-    aws_rds_aurora_show.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # aws-rds-aurora display sub-command option
-    aws_rds_aurora_display.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # aws-rds-aurora password sub-command option
-    aws_rds_aurora_passwords.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # aws-rds-aurora provision sub-command options
-    aws_rds_aurora_provision.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # aws-rds-aurora destroy sub-command options
-    aws_rds_aurora_destroy.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # aws-rds-aurora deploy sub-command options
-    aws_rds_aurora_deploy.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    aws_rds_aurora_deploy.add_argument(
-        '-n', '--no-install-collection',
-        dest='no_install_collection',
-        action='store_true',
-        help="Do not install the Ansible collection."
-    )
-
-# Azure sub-commands and options
-def azure_subcommands(azure_subparser):
-    azure_configure = azure_subparser.add_parser(
-        'configure', help='Project configuration'
-    )
-    azure_provision = azure_subparser.add_parser(
-        'provision', help='Machines provisionning'
-    )
-    azure_destroy = azure_subparser.add_parser(
-        'destroy', help='Machines destruction'
-    )
-    azure_deploy = azure_subparser.add_parser(
-        'deploy', help='Postgres deployment'
-    )
-    azure_show = azure_subparser.add_parser(
-        'show', help='Show configuration'
-    )
-    azure_display = azure_subparser.add_parser(
-        'display', help='Display project details'
-    )
-    azure_passwords = azure_subparser.add_parser(
-        'passwords', help='Display project passwords'
-    )
-    azure_list = azure_subparser.add_parser(
-        'list', help='List projects'
-    )
-    azure_specs = azure_subparser.add_parser(
-        'specs', help='Show Cloud default specifications'
-    )
-    azure_logs = azure_subparser.add_parser(
-        'logs', help='Show project logs'
-    )
-    azure_remove = azure_subparser.add_parser(
-        'remove', help='Remove project'
-    )
-    # azure configure sub-command options
-    azure_configure.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Terraform project name'
-    ).completer = project_name_completer
-    azure_configure.add_argument(
-        '-a', '--reference-architecture',
-        dest='reference_architecture',
-        choices=ReferenceArchitectureOption.choices,
-        default=ReferenceArchitectureOption.default,
-        metavar='<ref-arch-code>',
-        help=ReferenceArchitectureOption.help
-    )
-    azure_configure.add_argument(
-        '-u', '--edb-credentials',
-        dest='edb_credentials',
-        required=True,
-        type=EDBCredentialsType,
-        metavar='"<username>:<password>"',
-        help="EDB Packages repository credentials."
-    ).completer = edb_credentials_completer
-    azure_configure.add_argument(
-        '-o', '--os',
-        dest='operating_system',
-        choices=OSOption.choices,
-        default=OSOption.default,
-        metavar='<operating-system>',
-        help=OSOption.help
-    )
-    azure_configure.add_argument(
-        '-t', '--pg-type',
-        dest='postgres_type',
-        choices=PgTypeOption.choices,
-        default=PgTypeOption.default,
-        metavar='<postgres-engine-type>',
-        help=PgTypeOption.help
-    )
-    azure_configure.add_argument(
-        '-v', '--pg-version',
-        dest='postgres_version',
-        choices=PgVersionOption.choices,
-        default=PgVersionOption.default,
-        metavar='<postgres-version>',
-        help=PgVersionOption.help
-    )
-    azure_configure.add_argument(
-        '-e', '--efm-version',
-        dest='efm_version',
-        choices=EFMVersionOption.choices,
-        default=EFMVersionOption.default,
-        metavar='<efm-version>',
-        help=EFMVersionOption.help
-    )
-    azure_configure.add_argument(
-        '-k', '--ssh-pub-key',
-        dest='ssh_pub_key',
-        type=argparse.FileType('r'),
-        default=SSHPubKeyOption.default(),
-        metavar='<ssh-public-key-file>',
-        help=SSHPubKeyOption.help
-    )
-    azure_configure.add_argument(
-        '-K', '--ssh-private-key',
-        dest='ssh_priv_key',
-        type=argparse.FileType('r'),
-        default=SSHPrivKeyOption.default(),
-        metavar='<ssh-private-key-file>',
-        help=SSHPrivKeyOption.help
-    )
-    azure_configure.add_argument(
-        '-s', '--spec',
-        dest='spec_file',
-        type=argparse.FileType('r'),
-        metavar='<azure-spec-file>',
-        help="Azure instances specification file, in JSON."
-    )
-    azure_configure.add_argument(
-        '-r', '--azure-region',
-        dest='azure_region',
-        choices=AzureRegionOption.choices,
-        default=AzureRegionOption.default,
-        metavar='<cloud-region>',
-        help=AzureRegionOption.help
-    )
-    # azure logs sub-command options
-    azure_logs.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    azure_logs.add_argument(
-        '-t', '--tail',
-        dest='tail',
-        action='store_true',
-        help="Do not stop at the end of file."
-    )
-    # azure remove sub-command options
-    azure_remove.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # azure show sub-command options
-    azure_show.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # azure display sub-command options
-    azure_display.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # azure password sub-command options
-    azure_passwords.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # azure provision sub-command options
-    azure_provision.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # azure destroy sub-command options
-    azure_destroy.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # azure deploy sub-command options
-    azure_deploy.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    azure_deploy.add_argument(
-        '-n', '--no-install-collection',
-        dest='no_install_collection',
-        action='store_true',
-        help="Do not install the Ansible collection."
-    )
-
-# GCloud sub-commands and options
-def gcloud_subcommands(gcloud_subparser):
-    gcloud_configure = gcloud_subparser.add_parser(
-        'configure', help='Project configuration'
-    )
-    gcloud_provision = gcloud_subparser.add_parser(
-        'provision', help='Machines provisionning'
-    )
-    gcloud_destroy = gcloud_subparser.add_parser(
-        'destroy', help='Machines destruction'
-    )
-    gcloud_deploy = gcloud_subparser.add_parser(
-        'deploy', help='Postgres deployment'
-    )
-    gcloud_show = gcloud_subparser.add_parser(
-        'show', help='Show configuration'
-    )
-    gcloud_display = gcloud_subparser.add_parser(
-        'display', help='Display project details'
-    )
-    gcloud_passwords = gcloud_subparser.add_parser(
-        'passwords', help='Display project passwords'
-    )
-    gcloud_list = gcloud_subparser.add_parser(
-        'list', help='List projects'
-    )
-    gcloud_specs = gcloud_subparser.add_parser(
-        'specs', help='Show Cloud default specifications'
-    )
-    gcloud_logs = gcloud_subparser.add_parser(
-        'logs', help='Show project logs'
-    )
-    gcloud_remove = gcloud_subparser.add_parser(
-        'remove', help='Remove project'
-    )
-    # gcloud configure sub-command options
-    gcloud_configure.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Terraform project name'
-    ).completer = project_name_completer
-    gcloud_configure.add_argument(
-        '-a', '--reference-architecture',
-        dest='reference_architecture',
-        choices=ReferenceArchitectureOption.choices,
-        default=ReferenceArchitectureOption.default,
-        metavar='<ref-arch-code>',
-        help=ReferenceArchitectureOption.help
-    )
-    gcloud_configure.add_argument(
-        '-u', '--edb-credentials',
-        dest='edb_credentials',
-        required=True,
-        type=EDBCredentialsType,
-        metavar='"<username>:<password>"',
-        help="EDB Packages repository credentials."
-    ).completer = edb_credentials_completer
-    gcloud_configure.add_argument(
-        '-o', '--os',
-        dest='operating_system',
-        choices=OSOption.choices,
-        default=OSOption.default,
-        metavar='<operating-system>',
-        help=OSOption.help
-    )
-    gcloud_configure.add_argument(
-        '-t', '--pg-type',
-        dest='postgres_type',
-        choices=PgTypeOption.choices,
-        default=PgTypeOption.default,
-        metavar='<postgres-engine-type>',
-        help=PgTypeOption.help
-    )
-    gcloud_configure.add_argument(
-        '-v', '--pg-version',
-        dest='postgres_version',
-        choices=PgVersionOption.choices,
-        default=PgVersionOption.default,
-        metavar='<postgres-version>',
-        help=PgVersionOption.help
-    )
-    gcloud_configure.add_argument(
-        '-e', '--efm-version',
-        dest='efm_version',
-        choices=EFMVersionOption.choices,
-        default=EFMVersionOption.default,
-        metavar='<efm-version>',
-        help=EFMVersionOption.help
-    )
-    gcloud_configure.add_argument(
-        '-k', '--ssh-pub-key',
-        dest='ssh_pub_key',
-        type=argparse.FileType('r'),
-        default=SSHPubKeyOption.default(),
-        metavar='<ssh-public-key-file>',
-        help=SSHPubKeyOption.help
-    )
-    gcloud_configure.add_argument(
-        '-K', '--ssh-private-key',
-        dest='ssh_priv_key',
-        type=argparse.FileType('r'),
-        default=SSHPrivKeyOption.default(),
-        metavar='<ssh-private-key-file>',
-        help=SSHPrivKeyOption.help
-    )
-    gcloud_configure.add_argument(
-        '-r', '--gcloud-region',
-        dest='gcloud_region',
-        choices=GCloudRegionOption.choices,
-        default=GCloudRegionOption.default,
-        metavar='<cloud-region>',
-        help=GCloudRegionOption.help
-    )
-    gcloud_configure.add_argument(
-        '-s', '--spec',
-        dest='spec_file',
-        type=argparse.FileType('r'),
-        metavar='<gcloud-spec-file>',
-        help="GCloud instances specification file, in JSON."
-    )
-    gcloud_configure.add_argument(
-        '-c', '--gcloud-credentials',
-        dest='gcloud_credentials',
-        type=argparse.FileType('r'),
-        default=GCloudCredentialsOption.default(),
-        metavar='<gcloud-credentials-json-file>',
-        help=GCloudCredentialsOption.help
-    )
-    gcloud_configure.add_argument(
-        '-p', '--gcloud-project-id',
-        dest='gcloud_project_id',
-        required=True,
-        type=str,
-        metavar='<gcloud-project-id>',
-        help="GCloud project ID"
-    ).completer = gcloud_project_id_completer
-    # gcloud logs sub-command options
-    gcloud_logs.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    gcloud_logs.add_argument(
-        '-t', '--tail',
-        dest='tail',
-        action='store_true',
-        help="Do not stop at the end of file."
-    )
-    # gcloud remove sub-command options
-    gcloud_remove.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # gcloud show sub-command options
-    gcloud_show.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # gcloud display sub-command options
-    gcloud_display.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # gcloud password sub-command options
-    gcloud_passwords.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # gcloud provision sub-command options
-    gcloud_provision.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # gcloud destroy sub-command options
-    gcloud_destroy.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    # gcloud deploy sub-command options
-    gcloud_deploy.add_argument(
-        'project', type=ProjectType, metavar='<project-name>',
-        help='Project name'
-    ).completer = project_name_completer
-    gcloud_deploy.add_argument(
-        '-n', '--no-install-collection',
-        dest='no_install_collection',
-        action='store_true',
-        help="Do not install the Ansible collection."
-    )
-
-# Argcomplete completer
-def project_name_completer(prefix, parsed_args, **kwargs):
-    cloud = parsed_args.cloud
-    sub_command = parsed_args.sub_command
-    if sub_command == 'configure':
-        return ("PROJECT_NAME",)
-    else:
-        projects_path = os.path.join(Project.projects_root_path, cloud)
-        if not os.path.exists(projects_path):
-            return ("PROJECT_NAME",)
-        return (pname for pname in os.listdir(projects_path))
-
-def edb_credentials_completer(prefix, parsed_args, **kwargs):
-    return ("USERNAME:PASSWORD",)
-
-def aws_ami_id_completer(prefix, parsed_args, **kwargs):
-    return ("AWS_AMI_ID",)
-
-def gcloud_project_id_completer(prefix, parsed_args, **kwargs):
-    return ("GCLOUD_PROJECT_ID",)
 
 def parse():
     parser = CLIParser(
@@ -1135,41 +34,40 @@ def parse():
         title='Cloud provider', dest='cloud', metavar='<cloud>'
     )
 
-    # Cloud commands
-    aws = subparsers.add_parser('aws', help='AWS Cloud')
-    aws_rds = subparsers.add_parser('aws-rds', help='AWS RDS Cloud')
-    aws_rds_aurora = subparsers.add_parser('aws-rds-aurora', help='AWS RDS Aurora Cloud')
-    azure = subparsers.add_parser('azure', help='Azure Cloud')
-    gcloud = subparsers.add_parser('gcloud', help='Google Cloud')
+    # Cloud commands parsers
+    aws_parser = subparsers.add_parser('aws', help='AWS Cloud')
+    aws_rds_parser = subparsers.add_parser('aws-rds', help='AWS RDS Cloud')
+    aws_rds_aurora_parser = subparsers.add_parser(
+        'aws-rds-aurora', help='AWS RDS Aurora Cloud'
+    )
+    azure_parser = subparsers.add_parser('azure', help='Azure Cloud')
+    gcloud_parser = subparsers.add_parser('gcloud', help='Google Cloud')
 
-    # Sub-commands
-    # AWS
-    aws_subparser = aws.add_subparsers(
+    # Sub-commands parsers
+    aws_subparser = aws_parser.add_subparsers(
         title='AWS sub-commands', dest='sub_command', metavar='<sub-command>'
     )
-    aws_subcommands(aws_subparser)
-    # AWS RDS
-    aws_rds_subparser = aws_rds.add_subparsers(
+    aws_rds_subparser = aws_rds_parser.add_subparsers(
         title='AWS RDS sub-commands', dest='sub_command', metavar='<sub-command>'
     )
-    aws_rds_subcommands(aws_rds_subparser)
-    # AWS RDS Aurora
-    aws_rds_aurora_subparser = aws_rds_aurora.add_subparsers(
-        title='AWS RDS Aurora sub-commands', dest='sub_command', metavar='<sub-command>'
-    )
-    aws_rds_aurora_subcommands(aws_rds_aurora_subparser)
-    # Azure
-    azure_subparser = azure.add_subparsers(
-        title='Azure sub-commands', dest='sub_command', metavar='<sub-command>'
-    )
-    azure_subcommands(azure_subparser)
-    # GCloud
-    gcloud_subparser = gcloud.add_subparsers(
-        title='GCloud sub-commands',
-        dest='sub_command',
+    aws_rds_aurora_subparser = aws_rds_aurora_parser.add_subparsers(
+        title='AWS RDS Aurora sub-commands', dest='sub_command',
         metavar='<sub-command>'
     )
-    gcloud_subcommands(gcloud_subparser)
+    azure_subparser = azure_parser.add_subparsers(
+        title='Azure sub-commands', dest='sub_command', metavar='<sub-command>'
+    )
+    gcloud_subparser = gcloud_parser.add_subparsers(
+        title='GCloud sub-commands', dest='sub_command',
+        metavar='<sub-command>'
+    )
+
+    # Attach sub-commands options to the sub-parsers
+    aws.subcommands(aws_subparser)
+    aws_rds.subcommands(aws_rds_subparser)
+    aws_rds_aurora.subcommands(aws_rds_aurora_subparser)
+    azure.subcommands(azure_subparser)
+    gcloud.subcommands(gcloud_subparser)
 
     # Autocompletion with argcomplete
     argcomplete.autocomplete(parser)
@@ -1191,15 +89,15 @@ def parse():
     # Check if the <sub-command> argument is set
     if not getattr(env, 'sub_command'):
         if env.cloud == 'aws':
-            aws.print_help()
+            aws_parser.print_help()
         elif env.cloud == 'aws-rds':
-            aws_rds.print_help()
+            aws_rds_parser.print_help()
         elif env.cloud == 'aws-rds-aurora':
-            aws_rds_aurora.print_help()
+            aws_rds_aurora_parser.print_help()
         elif env.cloud == 'azure':
-            azure.print_help()
+            azure_parser.print_help()
         elif env.cloud == 'gcloud':
-            gcloud.print_help()
+            gcloud_parser.print_help()
         sys.stderr.write('error: too few arguments\n')
         sys.exit(2)
